Moved IWindowsManager impl to own namespace
